ASSIGNED - # 189: Fair Editor Events table is not updated when the underlying Premises document changes. http://localhost:8000/tracker/ticket/189
This is a Synchronization of Resources problem between Tracker, TrackerTransaction, and Fair editors all three should be able to save 
an underlying resource and the others should see that it has changed, reload inputs to tables, and reset the EContentAdapters to watch the 
new resource. 

This commit is close solving this.