ASSIGNED - # 289: Consolidate Tracker and Fair into two separate Feature Groups and clean up other features. http://localhost:8000/tracker/ticket/289
Refactored c.v.t.ocl.query classes to remove all references to Tracker and Fair models.  This bundle is now a container for generic classes that can be queried.  In the case of Context queries the Package URI will have to be passed into the delegate as a Parameterized Type.